Type of polynomial_order is changed from int to uint8_t. Tests, parameter checks, errors related to negative value are all removed. Refs #232
git-svn-id: http://alma-dms/svn/sakura/trunk@408 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/libsakura/optimized_implementation_factory.h
@@ -117,7 +117,7 @@ public:
 	}
 	virtual LIBSAKURA_SYMBOL(Status) Interpolate1D(
 	LIBSAKURA_SYMBOL(InterpolationMethod) interpolation_method,
-			int polynomial_order, size_t num_base,
+			uint8_t polynomial_order, size_t num_base,
 			XDataType const x_base[/*num_base*/], size_t num_array,
 			YDataType const y_base[/*num_base * num_base_array*/],
 			size_t num_interpolated,
